Allow immediate alpha setting when speed = 0; refactoring
* Separate fade logic into P_FindFakeFloorsDoAlpha, P_DoFakeFloorAlpha
* Change INT32 destvalue and speed to INT16

+++ src/p_spec.h
@@ -454,8 +454,8 @@ typedef struct
 {
 	thinker_t thinker;  ///< Thinker structure for effect.
 	INT32 affectee;     ///< Number of affected line
-	INT32 destvalue;    ///< Transparency value to fade to
-	INT32 speed;        ///< Speed to fade by
+	INT16 destvalue;    ///< Transparency value to fade to
+	INT16 speed;        ///< Speed to fade by
 	boolean doexists;      ///< Handle FF_EXISTS handling
 	boolean dotranslucent; ///< Handle FF_TRANSLUCENT handling
 	boolean dosolid;       ///< Handle FF_SOLID handling
